// frontend/src/client/types.gen.ts
export type IngestionTaskRead = {
    /**
     * Id
     */
    id: string;
    /**
     * Task Type
     */
    task_type: 'ingest' | 'reindex';
    /**
     * Status
     */
    status: 'pending' | 'running' | 'success' | 'failed';
    /**
     * Retry Count
     */
    retry_count: number;
    /**
     * Error Message
     */
    error_message?: string | null;
    /**
     * Progress Total
     */
    progress_total: number;
    /**
     * Progress Done
     */
    progress_done: number;
    /**
     * Started At
     */
    started_at?: string | null;
    /**
     * Finished At
     */
    finished_at?: string | null;
    /**
     * Created At
     */
    created_at: string;
};